ci/cd

+++ bin/unpublocal.sh
@@ -1,5 +1,19 @@
-# Remove temp directory and all its contents recursively if it exists
-rm -rf temp
+# Parse command line arguments
+CI=false
+
+while [[ "$#" -gt 0 ]]; do
+    case $1 in
+        --ci) CI=true ;;
+        *) echo "Unknown parameter: $1"; exit 1 ;;
+    esac
+    shift
+done
+
+if [ "$CI" = false ]; then
+    # Remove temp directory and all its contents recursively if it exists
+    rm -rf temp
+fi
+
 
 # Get content from packages.ts and replace current content
 cat > sdk/tests/packages.ts << 'EOL'
